+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVORecipeSeralizers.java
@@ -1,9 +1,6 @@
 package com.CuteNekoDragon.Core.common.data;
 
-import com.CuteNekoDragon.Core.common.recipe.serializers.SackToBackpackRecipe;
-import com.CuteNekoDragon.Core.common.recipe.serializers.SackToLunchboxRecipe;
-import com.CuteNekoDragon.Core.common.recipe.serializers.SackToToolbeltRecipe;
-import com.CuteNekoDragon.Core.common.recipe.serializers.SackUpgradeRecipeSerializer;
+import com.CuteNekoDragon.Core.common.recipe.serializers.*;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -21,4 +18,7 @@ public class SVORecipeSeralizers {
     public static final RegistryObject<SackUpgradeRecipeSerializer<SackToLunchboxRecipe>> SACK_TO_LUNCHBOX = RECIPE_SERIALIZERS.register("sack_to_lunchbox", () -> new SackUpgradeRecipeSerializer<>(SackToLunchboxRecipe::new));
 
     public static final RegistryObject<SackUpgradeRecipeSerializer<SackToToolbeltRecipe>> SACK_TO_TOOLBELT = RECIPE_SERIALIZERS.register("sack_to_toolbelt", () -> new SackUpgradeRecipeSerializer<>(SackToToolbeltRecipe::new));
+
+    public static final RegistryObject<RecipeSerializer<KeepContentsSmithingRecipe>> SMITHING_KEEP_CONTENTS =
+            RECIPE_SERIALIZERS.register("smithing_keep_contents", KeepContentsSmithingRecipe.Serializer::new);
 }