+++ changelog.md
@@ -3,6 +3,7 @@
 - fixed recipe to upgrade sack into lunchbox, now it properly makes items stay and refunds incompatible items.
 - fixed recipe to upgrade sack into toolbelt, now it properly makes items stay and refunds incompatible items.
 - fixed recipe to upgrade sack into backpacks, now it properly makes items stay and refunds incompatible items.
+- upgrade recipes for lunchbox and toolbelt keep items now.
 
 ## [0.0.33]
 ### Changes